+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.1.2
+* Add attribute for worker advertise address
+* Bind service to 0.0.0.0:2376
+* Additional testing and documentation
+
 ## 0.1.0
 Inital release